example 2 coming along

+++ engine/src/geometry/geometry.rs
@@ -1,22 +1,21 @@
 use std::sync::Mutex;
 
-use crate::math::linear_algebra::{mat4::Mat4, vec3::Vec3, vec4::Vec4};
+use crate::math::linear_algebra::vec3::Vec3;
 
 use super::bind_group::GeometryBindGroupObject;
 
-pub type GeometryId = u64;
+pub type GeometryId = u32;
 
-static mut GEOMETRY_ID_GENERATOR: Mutex<GeometryId> = Mutex::new(0u64);
+static mut GEOMETRY_ID_GENERATOR: Mutex<GeometryId> = Mutex::new(0u32);
 
 pub fn new_geometry_id() -> GeometryId {
     unsafe {
         let mut changer = GEOMETRY_ID_GENERATOR.lock().unwrap();
-        *changer += 1u64;
+        *changer += 1u32;
         *changer
     }
 }
 
-// TODO: time for polymorphism
 pub trait Geometry {
     fn get_bind_group_object_mut(&mut self) -> &mut GeometryBindGroupObject;
